Display chart after vote
Refs #6
